feat: restore support to ignore-types flag
